AzureMonitor: Remove unused code
