+++ public/app/plugins/datasource/grafana-azure-monitor-datasource/components/MetricsQueryEditor/MetricsQueryEditor.test.tsx
@@ -1,11 +1,20 @@
-import { render, screen, waitFor } from '@testing-library/react';
+import { render, screen } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
 import React from 'react';
 
+import { config } from '@grafana/runtime';
 import { selectOptionInTest } from '@grafana/ui';
 
 import createMockDatasource from '../../__mocks__/datasource';
+import { createMockInstanceSetttings } from '../../__mocks__/instanceSettings';
 import createMockPanelData from '../../__mocks__/panelData';
 import createMockQuery from '../../__mocks__/query';
+import {
+  createMockResourceGroupsBySubscription,
+  createMockSubscriptions,
+  mockResourcesByResourceGroup,
+} from '../../__mocks__/resourcePickerRows';
+import ResourcePickerData from '../../resourcePicker/resourcePickerData';
 
 import MetricsQueryEditor from './MetricsQueryEditor';
 
@@ -14,259 +23,238 @@ const variableOptionGroup = {
   options: [],
 };
 
-describe('Azure Monitor QueryEditor', () => {
-  const mockPanelData = createMockPanelData();
-  it('should render', async () => {
-    const mockDatasource = createMockDatasource();
-    render(
-      <MetricsQueryEditor
-        data={mockPanelData}
-        subscriptionId="123"
-        query={createMockQuery()}
-        datasource={mockDatasource}
-        variableOptionGroup={variableOptionGroup}
-        onChange={() => {}}
-        setError={() => {}}
-      />
-    );
-    await waitFor(() => expect(screen.getByTestId('azure-monitor-metrics-query-editor')).toBeInTheDocument());
-  });
+const tests = [
+  {
+    id: 'azure-monitor-metrics-query-editor-with-resource-picker',
+  },
+  {
+    id: 'azure-monitor-metrics-query-editor-with-experimental-ui',
+  },
+];
 
-  it('should change the subscription ID when selected', async () => {
-    const mockDatasource = createMockDatasource();
-    const onChange = jest.fn();
-    const mockQuery = createMockQuery();
-    (mockQuery.azureMonitor ?? {}).metricName = undefined;
-    mockDatasource.azureMonitorDatasource.getSubscriptions = jest.fn().mockResolvedValueOnce([
-      {
-        value: 'abc-123',
-        text: 'Primary Subscription',
-      },
-      {
-        value: 'abc-456',
-        text: 'Another Subscription',
-      },
-    ]);
+export function createMockResourcePickerData() {
+  const mockDatasource = new ResourcePickerData(createMockInstanceSetttings());
 
-    render(
-      <MetricsQueryEditor
-        data={mockPanelData}
-        subscriptionId="123"
-        query={mockQuery}
-        datasource={mockDatasource}
-        variableOptionGroup={variableOptionGroup}
-        onChange={onChange}
-        setError={() => {}}
-      />
-    );
+  mockDatasource.getSubscriptions = jest.fn().mockResolvedValue(createMockSubscriptions());
+  mockDatasource.getResourceGroupsBySubscriptionId = jest
+    .fn()
+    .mockResolvedValue(createMockResourceGroupsBySubscription());
+  mockDatasource.getResourcesForResourceGroup = jest.fn().mockResolvedValue(mockResourcesByResourceGroup());
+  mockDatasource.getResourceURIFromWorkspace = jest.fn().mockReturnValue('');
+  mockDatasource.getResourceURIDisplayProperties = jest.fn().mockResolvedValue({});
 
-    const subscriptions = await screen.findByLabelText('Subscription');
-    await selectOptionInTest(subscriptions, 'Another Subscription');
+  return mockDatasource;
+}
 
-    expect(onChange).toHaveBeenCalledWith({
-      ...mockQuery,
-      subscription: 'abc-456',
-      azureMonitor: {
-        ...mockQuery.azureMonitor,
-        resourceGroup: undefined,
-        metricDefinition: undefined,
-        metricNamespace: undefined,
-        resourceName: undefined,
-        metricName: undefined,
-        aggregation: undefined,
-        timeGrain: '',
-        dimensionFilters: [],
-      },
+for (const t of tests) {
+  describe(`MetricsQueryEditor: ${t.id}`, () => {
+    const originalScrollIntoView = window.HTMLElement.prototype.scrollIntoView;
+    const mockPanelData = createMockPanelData();
+
+    beforeEach(() => {
+      window.HTMLElement.prototype.scrollIntoView = function () {};
+      config.featureToggles.azureMonitorExperimentalUI =
+        t.id === 'azure-monitor-metrics-query-editor-with-experimental-ui';
+    });
+    afterEach(() => {
+      window.HTMLElement.prototype.scrollIntoView = originalScrollIntoView;
+      config.featureToggles.azureMonitorExperimentalUI = false;
+    });
+
+    it('should render', async () => {
+      const mockDatasource = createMockDatasource({ resourcePickerData: createMockResourcePickerData() });
+
+      render(
+        <MetricsQueryEditor
+          data={mockPanelData}
+          query={createMockQuery()}
+          datasource={mockDatasource}
+          variableOptionGroup={variableOptionGroup}
+          onChange={() => {}}
+          setError={() => {}}
+        />
+      );
+
+      expect(await screen.findByTestId(t.id)).toBeInTheDocument();
+    });
+
+    it('should change resource when a resource is selected in the ResourcePicker', async () => {
+      const mockDatasource = createMockDatasource({ resourcePickerData: createMockResourcePickerData() });
+      const query = createMockQuery();
+      delete query?.azureMonitor?.resourceUri;
+      const onChange = jest.fn();
+
+      render(
+        <MetricsQueryEditor
+          data={mockPanelData}
+          query={query}
+          datasource={mockDatasource}
+          variableOptionGroup={variableOptionGroup}
+          onChange={onChange}
+          setError={() => {}}
+        />
+      );
+
+      const resourcePickerButton = await screen.findByRole('button', { name: 'Select a resource' });
+      expect(resourcePickerButton).toBeInTheDocument();
+      expect(screen.queryByRole('button', { name: 'Expand Primary Subscription' })).not.toBeInTheDocument();
+      resourcePickerButton.click();
+
+      const subscriptionButton = await screen.findByRole('button', { name: 'Expand Primary Subscription' });
+      expect(subscriptionButton).toBeInTheDocument();
+      expect(screen.queryByRole('button', { name: 'Expand A Great Resource Group' })).not.toBeInTheDocument();
+      subscriptionButton.click();
+
+      const resourceGroupButton = await screen.findByRole('button', { name: 'Expand A Great Resource Group' });
+      expect(resourceGroupButton).toBeInTheDocument();
+      expect(screen.queryByLabelText('web-server')).not.toBeInTheDocument();
+      resourceGroupButton.click();
+
+      const checkbox = await screen.findByLabelText('web-server');
+      expect(checkbox).toBeInTheDocument();
+      expect(checkbox).not.toBeChecked();
+      await userEvent.click(checkbox);
+      expect(checkbox).toBeChecked();
+      await userEvent.click(await screen.findByRole('button', { name: 'Apply' }));
+
+      expect(onChange).toBeCalledTimes(1);
+      expect(onChange).toBeCalledWith(
+        expect.objectContaining({
+          azureMonitor: expect.objectContaining({
+            resourceUri:
+              '/subscriptions/def-456/resourceGroups/dev-3/providers/Microsoft.Compute/virtualMachines/web-server',
+          }),
+        })
+      );
+    });
+
+    it('should reset metric namespace, metric name, and aggregation fields after selecting a new resource when a valid query has already been set', async () => {
+      const mockDatasource = createMockDatasource({ resourcePickerData: createMockResourcePickerData() });
+      const query = createMockQuery();
+      const onChange = jest.fn();
+
+      render(
+        <MetricsQueryEditor
+          data={mockPanelData}
+          query={query}
+          datasource={mockDatasource}
+          variableOptionGroup={variableOptionGroup}
+          onChange={onChange}
+          setError={() => {}}
+        />
+      );
+
+      const resourcePickerButton = await screen.findByRole('button', { name: /grafana/ });
+
+      expect(screen.getByText('Microsoft.Compute/virtualMachines')).toBeInTheDocument();
+      expect(screen.getByText('Metric A')).toBeInTheDocument();
+      expect(screen.getByText('Average')).toBeInTheDocument();
+
+      expect(resourcePickerButton).toBeInTheDocument();
+      expect(screen.queryByRole('button', { name: 'Expand Primary Subscription' })).not.toBeInTheDocument();
+      resourcePickerButton.click();
+
+      const subscriptionButton = await screen.findByRole('button', { name: 'Expand Dev Subscription' });
+      expect(subscriptionButton).toBeInTheDocument();
+      expect(screen.queryByRole('button', { name: 'Expand Development 3' })).not.toBeInTheDocument();
+      subscriptionButton.click();
+
+      const resourceGroupButton = await screen.findByRole('button', { name: 'Expand Development 3' });
+      expect(resourceGroupButton).toBeInTheDocument();
+      expect(screen.queryByLabelText('db-server')).not.toBeInTheDocument();
+      resourceGroupButton.click();
+
+      const checkbox = await screen.findByLabelText('db-server');
+      expect(checkbox).toBeInTheDocument();
+      expect(checkbox).not.toBeChecked();
+      await userEvent.click(checkbox);
+      expect(checkbox).toBeChecked();
+      await userEvent.click(await screen.findByRole('button', { name: 'Apply' }));
+
+      expect(onChange).toBeCalledTimes(1);
+      expect(onChange).toBeCalledWith(
+        expect.objectContaining({
+          azureMonitor: expect.objectContaining({
+            resourceUri:
+              '/subscriptions/def-456/resourceGroups/dev-3/providers/Microsoft.Compute/virtualMachines/db-server',
+            metricNamespace: undefined,
+            metricName: undefined,
+            aggregation: undefined,
+            timeGrain: '',
+            dimensionFilters: [],
+          }),
+        })
+      );
+    });
+
+    it('should change the metric name when selected', async () => {
+      const mockDatasource = createMockDatasource({ resourcePickerData: createMockResourcePickerData() });
+      const onChange = jest.fn();
+      const mockQuery = createMockQuery();
+      mockDatasource.azureMonitorDatasource.getMetricNames = jest.fn().mockResolvedValue([
+        {
+          value: 'metric-a',
+          text: 'Metric A',
+        },
+        {
+          value: 'metric-b',
+          text: 'Metric B',
+        },
+      ]);
+
+      render(
+        <MetricsQueryEditor
+          data={mockPanelData}
+          query={createMockQuery()}
+          datasource={mockDatasource}
+          variableOptionGroup={variableOptionGroup}
+          onChange={onChange}
+          setError={() => {}}
+        />
+      );
+
+      const metrics = await screen.findByLabelText('Metric');
+      expect(metrics).toBeInTheDocument();
+      await selectOptionInTest(metrics, 'Metric B');
+
+      expect(onChange).toHaveBeenLastCalledWith({
+        ...mockQuery,
+        azureMonitor: {
+          ...mockQuery.azureMonitor,
+          metricName: 'metric-b',
+          aggregation: undefined,
+          timeGrain: '',
+        },
+      });
+    });
+
+    it('should change the aggregation type when selected', async () => {
+      const mockDatasource = createMockDatasource({ resourcePickerData: createMockResourcePickerData() });
+      const onChange = jest.fn();
+      const mockQuery = createMockQuery();
+
+      render(
+        <MetricsQueryEditor
+          data={mockPanelData}
+          query={createMockQuery()}
+          datasource={mockDatasource}
+          variableOptionGroup={variableOptionGroup}
+          onChange={onChange}
+          setError={() => {}}
+        />
+      );
+
+      const aggregation = await screen.findByLabelText('Aggregation');
+      expect(aggregation).toBeInTheDocument();
+      await selectOptionInTest(aggregation, 'Maximum');
+
+      expect(onChange).toHaveBeenLastCalledWith({
+        ...mockQuery,
+        azureMonitor: {
+          ...mockQuery.azureMonitor,
+          aggregation: 'Maximum',
+        },
+      });
     });
   });
-
-  it('should change the resource group when selected', async () => {
-    const mockDatasource = createMockDatasource();
-    const onChange = jest.fn();
-    const mockQuery = createMockQuery();
-    mockDatasource.getResourceGroups = jest.fn().mockResolvedValue([
-      { text: 'grafanastaging', value: 'grafanastaging' },
-      { text: 'Grafana Prod', value: 'grafanaprod' },
-    ]);
-    render(
-      <MetricsQueryEditor
-        data={mockPanelData}
-        subscriptionId="123"
-        query={createMockQuery()}
-        datasource={mockDatasource}
-        variableOptionGroup={variableOptionGroup}
-        onChange={onChange}
-        setError={() => {}}
-      />
-    );
-    await waitFor(() => expect(screen.getByTestId('azure-monitor-metrics-query-editor')).toBeInTheDocument());
-
-    const resourceGroup = await screen.findByLabelText('Resource group');
-    await selectOptionInTest(resourceGroup, 'Grafana Prod');
-
-    expect(onChange).toHaveBeenLastCalledWith({
-      ...mockQuery,
-      azureMonitor: {
-        ...mockQuery.azureMonitor,
-        resourceUri: '',
-        resourceGroup: 'grafanaprod',
-        metricDefinition: undefined,
-        metricNamespace: undefined,
-        resourceName: undefined,
-        metricName: undefined,
-        aggregation: undefined,
-        timeGrain: '',
-        dimensionFilters: [],
-      },
-    });
-  });
-
-  it('should change the resource type when selected', async () => {
-    const mockDatasource = createMockDatasource();
-    const onChange = jest.fn();
-    const mockQuery = createMockQuery();
-    mockDatasource.getMetricDefinitions = jest.fn().mockResolvedValue([
-      { text: 'Virtual Machine', value: 'azure/vm' },
-      { text: 'Database', value: 'azure/db' },
-    ]);
-    render(
-      <MetricsQueryEditor
-        data={mockPanelData}
-        subscriptionId="123"
-        query={createMockQuery()}
-        datasource={mockDatasource}
-        variableOptionGroup={variableOptionGroup}
-        onChange={onChange}
-        setError={() => {}}
-      />
-    );
-    await waitFor(() => expect(screen.getByTestId('azure-monitor-metrics-query-editor')).toBeInTheDocument());
-
-    const resourceGroup = await screen.findByLabelText('Resource type');
-    await selectOptionInTest(resourceGroup, 'Virtual Machine');
-
-    expect(onChange).toHaveBeenLastCalledWith({
-      ...mockQuery,
-      azureMonitor: {
-        ...mockQuery.azureMonitor,
-        resourceUri: '',
-        metricDefinition: 'azure/vm',
-        resourceName: undefined,
-        metricNamespace: undefined,
-        metricName: undefined,
-        aggregation: undefined,
-        timeGrain: '',
-        dimensionFilters: [],
-      },
-    });
-  });
-
-  it('should change the resource name when selected', async () => {
-    const mockDatasource = createMockDatasource();
-    const onChange = jest.fn();
-    const mockQuery = createMockQuery();
-    mockDatasource.getResourceNames = jest.fn().mockResolvedValue([
-      { text: 'ResourceName1', value: 'resource-name-1' },
-      { text: 'ResourceName2', value: 'resource-name-2' },
-    ]);
-    render(
-      <MetricsQueryEditor
-        data={mockPanelData}
-        subscriptionId="123"
-        query={createMockQuery()}
-        datasource={mockDatasource}
-        variableOptionGroup={variableOptionGroup}
-        onChange={onChange}
-        setError={() => {}}
-      />
-    );
-    await waitFor(() => expect(screen.getByTestId('azure-monitor-metrics-query-editor')).toBeInTheDocument());
-
-    const resourceGroup = await screen.findByLabelText('Resource name');
-    await selectOptionInTest(resourceGroup, 'ResourceName1');
-
-    expect(onChange).toHaveBeenLastCalledWith({
-      ...mockQuery,
-      azureMonitor: {
-        ...mockQuery.azureMonitor,
-        resourceUri: '',
-        resourceName: 'resource-name-1',
-        metricNamespace: undefined,
-        metricName: undefined,
-        aggregation: undefined,
-        timeGrain: '',
-        dimensionFilters: [],
-      },
-    });
-  });
-
-  it('should change the metric name when selected', async () => {
-    const mockDatasource = createMockDatasource();
-    const onChange = jest.fn();
-    const mockQuery = createMockQuery();
-    mockDatasource.azureMonitorDatasource.getMetricNames = jest.fn().mockResolvedValue([
-      {
-        value: 'metric-a',
-        text: 'Metric A',
-      },
-      {
-        value: 'metric-b',
-        text: 'Metric B',
-      },
-    ]);
-    render(
-      <MetricsQueryEditor
-        data={mockPanelData}
-        subscriptionId="123"
-        query={createMockQuery()}
-        datasource={mockDatasource}
-        variableOptionGroup={variableOptionGroup}
-        onChange={onChange}
-        setError={() => {}}
-      />
-    );
-    await waitFor(() => expect(screen.getByTestId('azure-monitor-metrics-query-editor')).toBeInTheDocument());
-
-    const metrics = await screen.findByLabelText('Metric');
-    await selectOptionInTest(metrics, 'Metric B');
-
-    expect(onChange).toHaveBeenLastCalledWith({
-      ...mockQuery,
-      azureMonitor: {
-        ...mockQuery.azureMonitor,
-        metricName: 'metric-b',
-        aggregation: undefined,
-        timeGrain: '',
-      },
-    });
-  });
-
-  it('should change the aggregation type when selected', async () => {
-    const mockDatasource = createMockDatasource();
-    const onChange = jest.fn();
-    const mockQuery = createMockQuery();
-    render(
-      <MetricsQueryEditor
-        data={mockPanelData}
-        subscriptionId="123"
-        query={createMockQuery()}
-        datasource={mockDatasource}
-        variableOptionGroup={variableOptionGroup}
-        onChange={onChange}
-        setError={() => {}}
-      />
-    );
-    await waitFor(() => expect(screen.getByTestId('azure-monitor-metrics-query-editor')).toBeInTheDocument());
-
-    const aggregation = await screen.findByLabelText('Aggregation');
-    await selectOptionInTest(aggregation, 'Maximum');
-
-    expect(onChange).toHaveBeenLastCalledWith({
-      ...mockQuery,
-      azureMonitor: {
-        ...mockQuery.azureMonitor,
-        aggregation: 'Maximum',
-      },
-    });
-  });
-});
+}
