+++ public/app/plugins/datasource/grafana-azure-monitor-datasource/components/MetricsQueryEditor/dataHooks.test.ts
@@ -5,17 +5,12 @@ import Datasource from '../../datasource';
 import { AzureMetricQuery, AzureMonitorOption, AzureMonitorQuery, AzureQueryType } from '../../types';
 
 import {
+  useMetricNames,
+  useMetricNamespaces,
+  useMetricMetadata,
   DataHook,
   MetricMetadata,
   MetricsMetadataHook,
-  updateSubscriptions,
-  useMetricMetadata,
-  useMetricNames,
-  useMetricNamespaces,
-  useResourceGroups,
-  useResourceNames,
-  useResourceTypes,
-  useSubscriptions,
 } from './dataHooks';
 
 const WAIT_OPTIONS = {
@@ -46,123 +41,16 @@ describe('AzureMonitor: metrics dataHooks', () => {
 
   const testTable: TestScenario[] = [
     {
-      name: 'useSubscriptions',
-      hook: useSubscriptions,
-      emptyQueryPartial: {},
-      topLevelCustomProperties: {
-        subscription: 'subscription-$ENVIRONMENT',
-      },
-      customProperties: {},
-      expectedOptions: [
-        {
-          label: 'sub-abc-123',
-          value: 'sub-abc-123',
-        },
-      ],
-      expectedCustomPropertyResults: [
-        {
-          label: 'sub-abc-123',
-          value: 'sub-abc-123',
-        },
-        {
-          label: 'subscription-$ENVIRONMENT',
-          value: 'subscription-$ENVIRONMENT',
-        },
-      ],
-    },
-
-    {
-      name: 'useResourceGroups',
-      hook: useResourceGroups,
-      emptyQueryPartial: {},
-      customProperties: {
-        resourceGroup: 'resource-group-$ENVIRONMENT',
-      },
-      expectedOptions: [
-        {
-          label: 'Web App - Production',
-          value: 'web-app-production',
-        },
-        {
-          label: 'Web App - Development',
-          value: 'web-app-development',
-        },
-      ],
-      expectedCustomPropertyResults: [
-        { label: 'Web App - Production', value: 'web-app-production' },
-        { label: 'Web App - Development', value: 'web-app-development' },
-        { label: 'resource-group-$ENVIRONMENT', value: 'resource-group-$ENVIRONMENT' },
-      ],
-    },
-
-    {
-      name: 'useResourceTypes',
-      hook: useResourceTypes,
-      emptyQueryPartial: {
-        resourceGroup: 'web-app-development',
-      },
-      customProperties: {
-        resourceGroup: 'web-app-development',
-        metricDefinition: 'azure/resource-type-$ENVIRONMENT',
-      },
-      expectedOptions: [
-        {
-          label: 'Virtual Machine',
-          value: 'azure/vm',
-        },
-        {
-          label: 'Database',
-          value: 'azure/db',
-        },
-      ],
-      expectedCustomPropertyResults: [
-        { label: 'Virtual Machine', value: 'azure/vm' },
-        { label: 'Database', value: 'azure/db' },
-        { label: 'azure/resource-type-$ENVIRONMENT', value: 'azure/resource-type-$ENVIRONMENT' },
-      ],
-    },
-    {
-      name: 'useResourceNames',
-      hook: useResourceNames,
-      emptyQueryPartial: {
-        resourceGroup: 'web-app-development',
-        metricDefinition: 'azure/vm',
-      },
-      customProperties: {
-        resourceGroup: 'web-app-development',
-        metricDefinition: 'azure/vm',
-        resourceName: 'resource-name-$ENVIRONMENT',
-      },
-      expectedOptions: [
-        {
-          label: 'Web server',
-          value: 'web-server',
-        },
-        {
-          label: 'Job server',
-          value: 'job-server',
-        },
-      ],
-      expectedCustomPropertyResults: [
-        { label: 'Web server', value: 'web-server' },
-        { label: 'Job server', value: 'job-server' },
-        { label: 'resource-name-$ENVIRONMENT', value: 'resource-name-$ENVIRONMENT' },
-      ],
-    },
-
-    {
       name: 'useMetricNames',
       hook: useMetricNames,
       emptyQueryPartial: {
-        resourceGroup: 'web-app-development',
-        metricDefinition: 'azure/vm',
-        resourceName: 'web-server',
+        resourceUri:
+          '/subscriptions/99999999-cccc-bbbb-aaaa-9106972f9572/resourceGroups/grafanastaging/providers/Microsoft.Compute/virtualMachines/grafana',
         metricNamespace: 'azure/vm',
       },
       customProperties: {
-        resourceGroup: 'web-app-development',
-        metricDefinition: 'azure/vm',
-        resourceName: 'web-server',
+        resourceUri:
+          '/subscriptions/99999999-cccc-bbbb-aaaa-9106972f9572/resourceGroups/grafanastaging/providers/Microsoft.Compute/virtualMachines/grafana',
         metricNamespace: 'azure/vm',
         metricName: 'metric-$ENVIRONMENT',
       },
@@ -182,20 +70,17 @@ describe('AzureMonitor: metrics dataHooks', () => {
         { label: 'metric-$ENVIRONMENT', value: 'metric-$ENVIRONMENT' },
       ],
     },
-
     {
       name: 'useMetricNamespaces',
       hook: useMetricNamespaces,
       emptyQueryPartial: {
-        resourceGroup: 'web-app-development',
-        metricDefinition: 'azure/vm',
-        resourceName: 'web-server',
+        resourceUri:
+          '/subscriptions/99999999-cccc-bbbb-aaaa-9106972f9572/resourceGroups/grafanastaging/providers/Microsoft.Compute/virtualMachines/grafana',
         metricNamespace: 'azure/vm',
       },
       customProperties: {
-        resourceGroup: 'web-app-development',
-        metricDefinition: 'azure/vm',
-        resourceName: 'web-server',
+        resourceUri:
+          '/subscriptions/99999999-cccc-bbbb-aaaa-9106972f9572/resourceGroups/grafanastaging/providers/Microsoft.Compute/virtualMachines/grafana',
         metricNamespace: 'azure/vm-$ENVIRONMENT',
         metricName: 'metric-name',
       },
@@ -307,11 +192,9 @@ describe('AzureMonitor: metrics dataHooks', () => {
       name: 'useMetricMetadata',
       hook: useMetricMetadata,
       emptyQueryPartial: {
-        resourceGroup: 'web-app-development',
-        metricDefinition: 'azure/vm',
-        resourceName: 'web-server',
+        resourceUri:
+          '/subscriptions/99999999-cccc-bbbb-aaaa-9106972f9572/resourceGroups/grafanastaging/providers/Microsoft.Compute/virtualMachines/grafana',
         metricNamespace: 'azure/vm',
-        subscription: 'test-sub',
         metricName: 'Average CPU',
       },
       customProperties: {},
@@ -353,84 +236,6 @@ describe('AzureMonitor: metrics dataHooks', () => {
           allowedTimeGrainsMs: [60_000, 300_000, 900_000, 1_800_000, 3_600_000, 21_600_000, 43_200_000, 86_400_000],
         },
       });
-    });
-  });
-});
-
-describe('AzureMonitor: updateSubscriptions', () => {
-  const bareQuery = {
-    refId: 'A',
-    queryType: AzureQueryType.AzureMonitor,
-  };
-
-  [
-    {
-      description: 'should not update with no subscriptions',
-      query: bareQuery,
-      subscriptionOptions: [],
-    },
-    {
-      description: 'should not update with the subscription as an option',
-      query: { ...bareQuery, subscription: 'foo' },
-      subscriptionOptions: [{ label: 'foo', value: 'foo' }],
-    },
-    {
-      description: 'should not update with a template variable',
-      query: { ...bareQuery, subscription: '$foo' },
-      subscriptionOptions: [],
-    },
-    {
-      description: 'should update with the first subscription',
-      query: { ...bareQuery },
-      subscriptionOptions: [{ label: 'foo', value: 'foo' }],
-      onChangeArgs: {
-        ...bareQuery,
-        subscription: 'foo',
-        azureMonitor: {
-          dimensionFilters: [],
-          timeGrain: '',
-          resourceUri: '',
-        },
-      },
-    },
-    {
-      description: 'should update with the default subscription if the current subsription does not exists',
-      query: { ...bareQuery, subscription: 'bar' },
-      subscriptionOptions: [{ label: 'foo', value: 'foo' }],
-      onChangeArgs: {
-        ...bareQuery,
-        subscription: 'foo',
-        azureMonitor: {
-          dimensionFilters: [],
-          timeGrain: '',
-          resourceUri: '',
-        },
-      },
-    },
-    {
-      description: 'should clean up if neither the default sub nor the current sub exists',
-      query: { ...bareQuery, subscription: 'bar' },
-      subscriptionOptions: [{ label: 'foo', value: 'foo' }],
-      defaultSubscription: 'foobar',
-      onChangeArgs: {
-        ...bareQuery,
-        subscription: '',
-        azureMonitor: {
-          dimensionFilters: [],
-          timeGrain: '',
-          resourceUri: '',
-        },
-      },
-    },
-  ].forEach((test) => {
-    it(test.description, () => {
-      const onChange = jest.fn();
-      updateSubscriptions(test.query, test.subscriptionOptions, onChange, test.defaultSubscription);
-      if (test.onChangeArgs) {
-        expect(onChange).toHaveBeenCalledWith(test.onChangeArgs);
-      } else {
-        expect(onChange).not.toHaveBeenCalled();
-      }
     });
   });
 });
