+++ public/app/plugins/datasource/grafana-azure-monitor-datasource/components/MetricsQueryEditor/dataHooks.ts
@@ -5,24 +5,12 @@ import { rangeUtil } from '@grafana/data';
 import Datasource from '../../datasource';
 import TimegrainConverter from '../../time_grain_converter';
 import { AzureMonitorErrorish, AzureMonitorOption, AzureMonitorQuery } from '../../types';
-import { hasOption, toOption } from '../../utils/common';
+import { toOption } from '../../utils/common';
 import { useAsyncState } from '../../utils/useAsyncState';
 
-import { setMetricNamespace, setSubscriptionID } from './setQueryValue';
-
-export interface MetricMetadata {
-  aggOptions: AzureMonitorOption[];
-  timeGrains: AzureMonitorOption[];
-  dimensions: AzureMonitorOption[];
-  isLoading: boolean;
-
-  // These two properties are only used within the hook, and not elsewhere
-  supportedAggTypes: string[];
-  primaryAggType: string | undefined;
-}
+import { setMetricNamespace } from './setQueryValue';
 
 type SetErrorFn = (source: string, error: AzureMonitorErrorish | undefined) => void;
-type OnChangeFn = (newQuery: AzureMonitorQuery) => void;
 
 export type DataHook = (
   query: AzureMonitorQuery,
@@ -37,138 +25,29 @@ export type MetricsMetadataHook = (
   onChange: OnChangeFn
 ) => MetricMetadata;
 
-export const updateSubscriptions = (
-  query: AzureMonitorQuery,
-  subscriptionOptions: AzureMonitorOption[],
-  onChange: OnChangeFn,
-  defaultSubscription?: string
-) => {
-  const { subscription } = query;
+export interface MetricMetadata {
+  aggOptions: AzureMonitorOption[];
+  timeGrains: AzureMonitorOption[];
+  dimensions: AzureMonitorOption[];
+  isLoading: boolean;
 
-  // Return early if subscriptions havent loaded, or if the query already has a subscription
-  if (
-    !subscriptionOptions.length ||
-    (subscription && (hasOption(subscriptionOptions, subscription) || subscription.includes('$')))
-  ) {
-    return;
-  }
+  // These two properties are only used within the hook, and not elsewhere
+  supportedAggTypes: string[];
+  primaryAggType: string | undefined;
+}
 
-  const defaultSub = defaultSubscription || subscriptionOptions[0].value;
-
-  if (!subscription && defaultSub && hasOption(subscriptionOptions, defaultSub)) {
-    onChange(setSubscriptionID(query, defaultSub));
-  }
-
-  // Check if the current subscription is in the list of subscriptions
-  if (subscription && !hasOption(subscriptionOptions, subscription)) {
-    if (hasOption(subscriptionOptions, defaultSub)) {
-      // Use the default sub if is on theh list
-      onChange(setSubscriptionID(query, defaultSub));
-    } else {
-      // Neither the current subscription nor the defaultSub is on the list, remove it
-      onChange(setSubscriptionID(query, ''));
-    }
-  }
-};
-
-export const useSubscriptions: DataHook = (query, datasource, onChange, setError) => {
-  const defaultSubscription = datasource.azureMonitorDatasource.defaultSubscriptionId;
-  const { subscription } = query;
-
-  const subscriptionOptions = useAsyncState(
-    async () => {
-      const results = await datasource.azureMonitorDatasource.getSubscriptions();
-      const options = formatOptions(results, subscription);
-
-      return options;
-    },
-    setError,
-    []
-  );
-
-  useEffect(() => {
-    updateSubscriptions(query, subscriptionOptions, onChange, defaultSubscription);
-  }, [subscriptionOptions, query, defaultSubscription, onChange]);
-
-  return subscriptionOptions;
-};
-
-export const useResourceGroups: DataHook = (query, datasource, onChange, setError) => {
-  const { subscription } = query;
-  const { resourceGroup } = query.azureMonitor ?? {};
-
-  return useAsyncState(
-    async () => {
-      if (!subscription) {
-        return;
-      }
-
-      const results = await datasource.getResourceGroups(subscription);
-      const options = formatOptions(results, resourceGroup);
-
-      return options;
-    },
-    setError,
-    [subscription]
-  );
-};
-
-export const useResourceTypes: DataHook = (query, datasource, onChange, setError) => {
-  const { subscription } = query;
-  const { resourceGroup, metricDefinition } = query.azureMonitor ?? {};
-
-  return useAsyncState(
-    async () => {
-      if (!(subscription && resourceGroup)) {
-        return;
-      }
-
-      const results = await datasource.getMetricDefinitions(subscription, resourceGroup);
-      const options = formatOptions(results, metricDefinition);
-
-      return options;
-    },
-    setError,
-    [subscription, resourceGroup]
-  );
-};
-
-export const useResourceNames: DataHook = (query, datasource, onChange, setError) => {
-  const { subscription } = query;
-  const { resourceGroup, metricDefinition, resourceName } = query.azureMonitor ?? {};
-
-  return useAsyncState(
-    async () => {
-      if (!(subscription && resourceGroup && metricDefinition)) {
-        return;
-      }
-
-      const results = await datasource.getResourceNames(subscription, resourceGroup, metricDefinition);
-      const options = formatOptions(results, resourceName);
-
-      return options;
-    },
-    setError,
-    [subscription, resourceGroup, metricDefinition]
-  );
-};
+type OnChangeFn = (newQuery: AzureMonitorQuery) => void;
 
 export const useMetricNamespaces: DataHook = (query, datasource, onChange, setError) => {
-  const { subscription } = query;
-  const { resourceGroup, metricDefinition, resourceName, metricNamespace } = query.azureMonitor ?? {};
+  const { metricNamespace, resourceUri } = query.azureMonitor ?? {};
 
   const metricNamespaces = useAsyncState(
     async () => {
-      if (!(subscription && resourceGroup && metricDefinition && resourceName)) {
+      if (!resourceUri) {
         return;
       }
 
-      const results = await datasource.azureMonitorDatasource.getMetricNamespaces({
-        subscription,
-        resourceGroup,
-        metricDefinition,
-        resourceName,
-      });
+      const results = await datasource.azureMonitorDatasource.getMetricNamespaces({ resourceUri });
       const options = formatOptions(results, metricNamespace);
 
       // Do some cleanup of the query state if need be
@@ -179,67 +58,61 @@ export const useMetricNamespaces: DataHook = (query, datasource, onChange, setEr
       return options;
     },
     setError,
-    [subscription, resourceGroup, metricDefinition, resourceName]
+    [resourceUri]
   );
 
   return metricNamespaces;
 };
 
 export const useMetricNames: DataHook = (query, datasource, onChange, setError) => {
-  const { subscription } = query;
-  const { resourceGroup, metricDefinition, resourceName, metricNamespace, metricName } = query.azureMonitor ?? {};
+  const { metricNamespace, metricName, resourceUri } = query.azureMonitor ?? {};
 
   return useAsyncState(
     async () => {
-      if (!(subscription && resourceGroup && metricDefinition && resourceName && metricNamespace)) {
+      if (!(metricNamespace && resourceUri)) {
         return;
       }
 
-      const results = await datasource.azureMonitorDatasource.getMetricNames({
-        subscription,
-        resourceGroup,
-        metricDefinition,
-        resourceName,
-        metricNamespace,
-      });
-
+      const results = await datasource.azureMonitorDatasource.getMetricNames({ resourceUri, metricNamespace });
       const options = formatOptions(results, metricName);
 
       return options;
     },
     setError,
-    [subscription, resourceGroup, metricDefinition, resourceName, metricNamespace]
+    [resourceUri, metricNamespace]
   );
 };
 
-export const useMetricMetadata = (query: AzureMonitorQuery, datasource: Datasource, onChange: OnChangeFn) => {
-  const [metricMetadata, setMetricMetadata] = useState<MetricMetadata>({
-    aggOptions: [],
-    timeGrains: [],
-    dimensions: [],
-    isLoading: false,
-    supportedAggTypes: [],
-    primaryAggType: undefined,
-  });
+const defaultMetricMetadata: MetricMetadata = {
+  aggOptions: [],
+  timeGrains: [],
+  dimensions: [],
+  isLoading: false,
+  supportedAggTypes: [],
+  primaryAggType: undefined,
+};
 
-  const { subscription } = query;
-  const { resourceGroup, metricDefinition, resourceName, metricNamespace, metricName, aggregation, timeGrain } =
-    query.azureMonitor ?? {};
+export const useMetricMetadata = (query: AzureMonitorQuery, datasource: Datasource, onChange: OnChangeFn) => {
+  const [metricMetadata, setMetricMetadata] = useState<MetricMetadata>(defaultMetricMetadata);
+
+  const { resourceUri, metricNamespace, metricName, aggregation, timeGrain } = query.azureMonitor ?? {};
 
   // Fetch new metric metadata when the fields change
   useEffect(() => {
-    if (!(subscription && resourceGroup && metricDefinition && resourceName && metricNamespace && metricName)) {
+    if (!(resourceUri && metricNamespace && metricName)) {
+      setMetricMetadata(defaultMetricMetadata);
       return;
     }
 
     datasource.azureMonitorDatasource
-      .getMetricMetadata({ subscription, resourceGroup, metricDefinition, resourceName, metricNamespace, metricName })
+      .getMetricMetadata({ resourceUri, metricNamespace, metricName })
       .then((metadata) => {
         // TODO: Move the aggregationTypes and timeGrain defaults into `getMetricMetadata`
         const aggregations = (metadata.supportedAggTypes || [metadata.primaryAggType]).map((v) => ({
           label: v,
           value: v,
         }));
+
         setMetricMetadata({
           aggOptions: aggregations,
           timeGrains: metadata.supportedTimeGrains,
@@ -249,7 +122,7 @@ export const useMetricMetadata = (query: AzureMonitorQuery, datasource: Datasour
           primaryAggType: metadata.primaryAggType,
         });
       });
-  }, [datasource, subscription, resourceGroup, metricDefinition, resourceName, metricNamespace, metricName]);
+  }, [datasource, resourceUri, metricNamespace, metricName]);
 
   // Update the query state in response to the meta data changing
   useEffect(() => {
@@ -286,7 +159,7 @@ function formatOptions(
   const options = rawResults.map(toOption);
 
   // account for custom values that might have been set in json file like ones crafted with a template variable (ex: "cloud-datasource-resource-$Environment")
-  if (selectedValue && !options.find((option) => option.value === selectedValue)) {
+  if (selectedValue && !options.find((option) => option.value === selectedValue.toLowerCase())) {
     options.push({ label: selectedValue, value: selectedValue });
   }
