--- public/app/plugins/datasource/grafana-azure-monitor-datasource/components/MetricsQueryEditor/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './MetricsQueryEditor';
