+++ public/app/plugins/datasource/grafana-azure-monitor-datasource/components/MetricsQueryEditor/setQueryValue.ts
@@ -15,95 +15,6 @@ export function setResource(query: AzureMonitorQuery, resourceURI: string | unde
   };
 }
 
-export function setSubscriptionID(query: AzureMonitorQuery, subscriptionID: string): AzureMonitorQuery {
-  if (query.subscription === subscriptionID) {
-    return query;
-  }
-
-  return {
-    ...query,
-    subscription: subscriptionID,
-    azureMonitor: {
-      ...query.azureMonitor,
-      resourceUri: '',
-      resourceGroup: undefined,
-      metricDefinition: undefined,
-      metricNamespace: undefined,
-      resourceName: undefined,
-      metricName: undefined,
-      aggregation: undefined,
-      timeGrain: '',
-      dimensionFilters: [],
-    },
-  };
-}
-
-export function setResourceGroup(query: AzureMonitorQuery, resourceGroup: string | undefined): AzureMonitorQuery {
-  if (query.azureMonitor?.resourceGroup === resourceGroup) {
-    return query;
-  }
-
-  return {
-    ...query,
-    azureMonitor: {
-      ...query.azureMonitor,
-      resourceUri: '',
-      resourceGroup: resourceGroup,
-      metricDefinition: undefined,
-      metricNamespace: undefined,
-      resourceName: undefined,
-      metricName: undefined,
-      aggregation: undefined,
-      timeGrain: '',
-      dimensionFilters: [],
-    },
-  };
-}
-
-// In the query as "metricDefinition" for some reason
-export function setResourceType(query: AzureMonitorQuery, resourceType: string | undefined): AzureMonitorQuery {
-  if (query.azureMonitor?.metricDefinition === resourceType) {
-    return query;
-  }
-
-  const newQuery = {
-    ...query,
-    azureMonitor: {
-      ...query.azureMonitor,
-      resourceUri: '',
-      metricDefinition: resourceType,
-      resourceName: undefined,
-      metricNamespace: undefined,
-      metricName: undefined,
-      aggregation: undefined,
-      timeGrain: '',
-      dimensionFilters: [],
-    },
-  };
-
-  return newQuery;
-}
-
-export function setResourceName(query: AzureMonitorQuery, resourceName: string | undefined): AzureMonitorQuery {
-  if (query.azureMonitor?.resourceName === resourceName) {
-    return query;
-  }
-
-  return {
-    ...query,
-    azureMonitor: {
-      ...query.azureMonitor,
-      resourceUri: '',
-      resourceName: resourceName,
-      metricNamespace: undefined,
-      metricName: undefined,
-      aggregation: undefined,
-      timeGrain: '',
-      dimensionFilters: [],
-    },
-  };
-}
-
 export function setMetricNamespace(query: AzureMonitorQuery, metricNamespace: string | undefined): AzureMonitorQuery {
   if (query.azureMonitor?.metricNamespace === metricNamespace) {
     return query;
